requirements.txt
Flask
Flask-Cors
openai
python-dotenv
streamlit
